+++ tools/Assembler/Model/Components/AluOperand.cs
@@ -6,7 +6,7 @@ public record AluOperand(
 	ConstantAst? Value
 ) : IAssemblyAst {
 	public AluOperand(CpuRegister register) : this(true, new CpuRegisterAst(register), null) {}
-	public AluOperand(short constantValue) : this(false, null, new ConstantAst(constantValue)) {}
+	public AluOperand(long constantValue) : this(false, null, new ConstantAst(constantValue)) {}
 	
 	public override string ToString() => IsRegister ? Register!.ToString() : Value!.ToString();
 }
