+++ tools/Assembler/Model/Components/ConstantAst.cs
@@ -1,5 +1,5 @@
 namespace Assembler;
 
-public record ConstantAst(short Value) : IAssemblyAst {
+public record ConstantAst(long Value) : IAssemblyAst {
 	public override string ToString() => $"constant {Value}";
 }
